# backend/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "WhatsApp Notification API",
    "version": "1.0.0",
    "description": "API for sending automated WhatsApp messages for booking and rent-out confirmations",
    "contact": {
      "name": "API Support",
      "email": "support@example.com"
    },
    "license": {
      "name": "ISC",
      "url": "https://opensource.org/licenses/ISC"
    }
  },
  "servers": [
    {
      "url": "http://localhost:5000",
      "description": "Development server"
    },
    {
      "url": "https://api.example.com",
      "description": "Production server"
    }
  ],
  "components": {
    "schemas": {
      "SendMessageRequest": {
        "type": "object",
        "required": [
          "brand",
          "event_type",
          "template_type",
          "customer_name",
          "customer_phone",
          "booking_number",
          "total_amount",
          "advance_paid",
          "balance_due"
        ],
        "properties": {
          "brand": {
            "type": "string",
            "enum": [
              "suitorguy",
              "zorucci"
            ],
            "description": "Brand name - SuitorGuy or Zorucci",
            "example": "suitorguy"
          },
          "event_type": {
            "type": "string",
            "enum": [
              "booking",
              "rentout"
            ],
            "description": "Type of event - booking or rentout",
            "example": "booking"
          },
          "template_type": {
            "type": "string",
            "enum": [
              "withdiscount",
              "nodisc"
            ],
            "description": "Template type - with discount or without discount",
            "example": "nodisc"
          },
          "customer_name": {
            "type": "string",
            "description": "Customer full name",
            "example": "John Doe"
          },
          "customer_phone": {
            "type": "string",
            "description": "Customer phone number with country code (no + sign)",
            "example": "918590292642"
          },
          "booking_number": {
            "type": "string",
            "description": "Unique booking or rent-out number",
            "example": "BK12345"
          },
          "total_amount": {
            "type": "string",
            "description": "Total amount in rupees",
            "example": "5000"
          },
          "discount_amount": {
            "type": "string",
            "description": "Discount amount (required for withdiscount template)",
            "example": "500"
          },
          "payable_amount": {
            "type": "string",
            "description": "Amount payable after discount",
            "example": "4500"
          },
          "invoice_amount": {
            "type": "string",
            "description": "Invoice amount (for rentout events)",
            "example": "7200"
          },
          "advance_paid": {
            "type": "string",
            "description": "Advance amount paid",
            "example": "2000"
          },
          "balance_due": {
            "type": "string",
            "description": "Remaining balance to be paid",
            "example": "2500"
          },
          "security_deposit": {
            "type": "string",
            "description": "Security deposit amount (for rentout with discount)",
            "example": "1000"
          },
          "security_amount": {
            "type": "string",
            "description": "Security amount (for rentout without discount)",
            "example": "1000"
          },
          "subtotal": {
            "type": "string",
            "description": "Subtotal amount (for rentout events)",
            "example": "8200"
          }
        }
      },
      "SuccessResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "message": {
            "type": "string",
            "example": "WhatsApp message sent successfully"
          },
          "data": {
            "type": "object",
            "properties": {
              "success": {
                "type": "boolean",
                "example": true
              },
              "messageId": {
                "type": "string",
                "description": "WhatsApp message ID from Meta API",
                "example": "wamid.HBgMOTE4NTkwMjkyNjQyFQIAERgSOTM4QTQ0MjFENzA4OUM2RTIzAA=="
              },
              "bookingNumber": {
                "type": "string",
                "example": "BK12345"
              }
            }
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": false
          },
          "message": {
            "type": "string",
            "example": "Failed to send WhatsApp message"
          },
          "error": {
            "type": "string",
            "example": "Error message details"
          },
          "errors": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "msg": {
                  "type": "string"
                },
                "param": {
                  "type": "string"
                },
                "location": {
                  "type": "string"
                }
              }
            }
          }
        }
      },
      "HealthResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "WhatsApp Notification Backend API"
          },
          "version": {
            "type": "string",
            "example": "1.0.0"
          },
          "status": {
            "type": "string",
            "example": "running"
          }
        }
      }
    }
  },
  "tags": [
    {
      "name": "Health",
      "description": "Health check endpoints"
    },
    {
      "name": "WhatsApp",
      "description": "WhatsApp message sending endpoints"
    }
  ],
  "paths": {
    "/whatsapp/send": {
      "post": {
        "summary": "Send WhatsApp notification message",
        "description": "Sends a WhatsApp template message to a customer for booking or rent-out confirmation",
        "tags": [
          "WhatsApp"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SendMessageRequest"
              },
              "examples": {
                "booking_without_discount": {
                  "summary": "Booking without discount",
                  "value": {
                    "brand": "suitorguy",
                    "event_type": "booking",
                    "template_type": "nodisc",
                    "customer_name": "John Doe",
                    "customer_phone": "918590292642",
                    "booking_number": "BK12345",
                    "total_amount": "5000",
                    "discount_amount": "0",
                    "payable_amount": "5000",
                    "advance_paid": "2000",
                    "balance_due": "3000"
                  }
                },
                "booking_with_discount": {
                  "summary": "Booking with discount",
                  "value": {
                    "brand": "suitorguy",
                    "event_type": "booking",
                    "template_type": "withdiscount",
                    "customer_name": "Jane Smith",
                    "customer_phone": "919876543210",
                    "booking_number": "BK12346",
                    "total_amount": "5000",
                    "discount_amount": "500",
                    "payable_amount": "4500",
                    "advance_paid": "2000",
                    "balance_due": "2500"
                  }
                },
                "rentout_with_discount": {
                  "summary": "Rent-out with discount",
                  "value": {
                    "brand": "zorucci",
                    "event_type": "rentout",
                    "template_type": "withdiscount",
                    "customer_name": "Mike Johnson",
                    "customer_phone": "919876543211",
                    "booking_number": "RO001",
                    "total_amount": "8000",
                    "discount_amount": "800",
                    "invoice_amount": "7200",
                    "advance_paid": "3000",
                    "balance_due": "4200",
                    "security_deposit": "1000",
                    "subtotal": "8200"
                  }
                },
                "rentout_without_discount": {
                  "summary": "Rent-out without discount",
                  "value": {
                    "brand": "zorucci",
                    "event_type": "rentout",
                    "template_type": "nodisc",
                    "customer_name": "Sarah Williams",
                    "customer_phone": "919876543212",
                    "booking_number": "RO002",
                    "total_amount": "8000",
                    "invoice_amount": "8000",
                    "advance_paid": "3000",
                    "balance_due": "5000",
                    "security_amount": "1000",
                    "subtotal": "9000"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Message sent successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                },
                "example": {
                  "success": true,
                  "message": "WhatsApp message sent successfully",
                  "data": {
                    "success": true,
                    "messageId": "wamid.HBgMOTE4NTkwMjkyNjQyFQIAERgSOTM4QTQ0MjFENzA4OUM2RTIzAA==",
                    "bookingNumber": "BK12345"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request - validation error or duplicate message",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "examples": {
                  "validation_error": {
                    "value": {
                      "success": false,
                      "errors": [
                        {
                          "msg": "Invalid brand",
                          "param": "brand",
                          "location": "body"
                        }
                      ]
                    }
                  },
                  "duplicate_message": {
                    "value": {
                      "success": false,
                      "message": "Message already sent for booking BK12345"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "success": false,
                  "message": "Failed to send WhatsApp message",
                  "error": "Invalid OAuth access token"
                }
              }
            }
          }
        }
      }
    },
    "/": {
      "get": {
        "summary": "Health check endpoint",
        "description": "Returns API status and version information",
        "tags": [
          "Health"
        ],
        "responses": {
          "200": {
            "description": "API is running",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HealthResponse"
                },
                "example": {
                  "message": "WhatsApp Notification Backend API",
                  "version": "1.0.0",
                  "status": "running"
                }
              }
            }
          }
        }
      }
    }
  }
}